@@ -16,7 +16,7 @@
   "author": "Maciej Krüger <mkg20001@gmail.com>",
   "license": "GPL-3.0",
   "dependencies": {
-    "electron": "^1.6.6",
+    "electron": "^2.0.9",
     "json-file": "^0.1.0"
   },
   "devDependencies": {},
