@@ -1,14 +1,14 @@
-"use strict"
+'use strict'
 
-const n = _ => _ < 10 ? _.toString() + "0" : _.toString()
+const n = _ => _ < 10 ? _.toString() + '0' : _.toString()
 
-function getWeekDay() {
-  return new Date().toLocaleString("en", {
+function getWeekDay () {
+  return new Date().toLocaleString('en', {
     weekday: 'long'
   }).toLowerCase()
 }
 
-function getDate() {
+function getDate () {
   const d = new Date()
   return n(d.getDate()) + n(d.getMonth()) + n(d.getFullYear())
 }
@@ -32,84 +32,84 @@ const timeTable = {
   }
 }
 
-function gotTime(time) {
+function gotTime (time) {
   const cur = new Date().getHours() * 100 + new Date().getMinutes()
-  return Object.keys(timeTable).filter(k => timeTable[k].from <= cur && timeTable[k].to >= cur).indexOf(time) != -1
+  return Object.keys(timeTable).filter(k => timeTable[k].from <= cur && timeTable[k].to >= cur).indexOf(time) !== -1
 }
 
 const types = {
   simple: {
     active: () => true,
-    tick: (task) => task.set("ticked", true),
-    untick: (task) => task.set("ticked", false),
-    state: (task) => !!task.get("ticked"),
+    tick: (task) => task.set('ticked', true),
+    untick: (task) => task.set('ticked', false),
+    state: (task) => Boolean(task.get('ticked')),
     render: (task) => task.text
   },
   days: {
     active: (task) => {
-      if (task.opt.days.indexOf("*") != -1) return true
+      if (task.opt.days.indexOf('*') !== -1) return true
       let l = getWeekDay()
       while (l.length > 1) {
-        if (task.opt.days.indexOf(l) != -1) return true
+        if (task.opt.days.indexOf(l) !== -1) return true
         l = l.substr(0, l.length - 1)
       }
       return false
     },
-    tick: (task) => task.set("ticked", getDate()),
-    untick: (task) => task.set("ticked", 0),
-    state: (task) => task.get("ticked") == getDate(),
+    tick: (task) => task.set('ticked', getDate()),
+    untick: (task) => task.set('ticked', 0),
+    state: (task) => task.get('ticked') === getDate(),
     render: (task) => task.text
   },
   daystime: {
     active: (task) => {
       if (!task.opt.times.filter(gotTime).length) return false
-      if (task.opt.days.indexOf("*") != -1) return true
+      if (task.opt.days.indexOf('*') !== -1) return true
       let l = getWeekDay()
       while (l.length > 1) {
-        if (task.opt.days.indexOf(l) != -1) return true
+        if (task.opt.days.indexOf(l) !== -1) return true
         l = l.substr(0, l.length - 1)
       }
       return false
     },
-    tick: (task) => task.set("ticked", getDate()),
-    untick: (task) => task.set("ticked", 0),
-    state: (task) => task.get("ticked") == getDate(),
+    tick: (task) => task.set('ticked', getDate()),
+    untick: (task) => task.set('ticked', 0),
+    state: (task) => task.get('ticked') === getDate(),
     render: (task) => task.text
   },
   time: {
-    active: (task) => !!task.opt.times.filter(gotTime).length,
-    tick: (task) => task.set("ticked", true),
-    untick: (task) => task.set("ticked", false),
-    state: (task) => !!task.get("ticked"),
+    active: (task) => Boolean(task.opt.times.filter(gotTime).length),
+    tick: (task) => task.set('ticked', true),
+    untick: (task) => task.set('ticked', false),
+    state: (task) => Boolean(task.get('ticked')),
     render: (task) => task.text
   }
 }
 
-function Task(data, db) {
-  //console.log(data, db)
+function Task (data, db) {
+  // console.log(data, db)
   const id = data.id
 
-  if (!db.get(id)) db.set(id, {}) //init
+  if (!db.get(id)) db.set(id, {}) // init
   const type = types[data.type]
   const self = this
-  if (!type) throw new Error("Unsupported type " + data.type)
+  if (!type) throw new Error('Unsupported type ' + data.type)
 
-  function wrap(prop) {
+  function wrap (prop) {
     return function () {
       var a = [].slice.call(arguments, 0)
       a.unshift(_task)
       return type[prop].apply(_task, a)
     }
   }
-  ["active", "tick", "untick", "state", "render"].map(s => self[s] = wrap(s));
-  ["group", "groupId", "id", "longterm"].map(s => self[s] = data[s])
+  ['active', 'tick', 'untick', 'state', 'render'].map(s => (self[s] = wrap(s)));
+  ['group', 'groupId', 'id', 'longterm'].map(s => (self[s] = data[s]))
 
-  function get(prop) {
-    return db.get(id + "." + prop)
+  function get (prop) {
+    return db.get(id + '.' + prop)
   }
 
-  function set(prop, val) {
-    db.set(id + "." + prop, val)
+  function set (prop, val) {
+    db.set(id + '.' + prop, val)
     db.writeSync()
   }
   const _task = {
@@ -117,6 +117,6 @@ function Task(data, db) {
     set: set,
     opt: data,
     text: data.text
-  } //mini object to pass to type
+  } // mini object to pass to type
 }
 module.exports = Task
