@@ -1,38 +1,38 @@
-"use strict"
+'use strict'
 
-const jf = require("json-file")
+const jf = require('json-file')
 
-const Task = require(__dirname + "/task")
+const Task = require(__dirname + '/task')
 
-function TaskDB(file) {
+function TaskDB (file) {
   const db = jf.read(file)
 
-  function save() {
+  function save () {
     db.writeSync()
   }
   let nt = []
   let nt_ = {}
   let gr = []
 
-  function tasks(_nt) {
+  function tasks (_nt) {
     nt_ = {}
-    //console.log(_nt)
+    // console.log(_nt)
     nt = _nt.map(n => new Task(n, db))
     let g = {}
-    nt.map(n => nt_[n.id] = n && (g[n.groupId] = {
+    nt.map(n => (nt_[n.id] = n && (g[n.groupId] = {
       id: n.groupId,
       g: n.group
-    }))
+    })))
     gr = Object.keys(g).map(k => g[k])
     save()
   }
 
-  function get(id) {
+  function get (id) {
     if (id) return nt_[id]
     else return nt
   }
 
-  function getG() {
+  function getG () {
     return gr
   }
   this.save = save
